requirements.txt
flask>=2.2.3
flask-login>=0.6.2
flask-SQLAlchemy>=3.0.3
argon2-cffi>=23.1.0
